Added a ball that reacts to being shot. We can tweak the mass of the bullets, the drag and bounce ratio of the ball, to get things working the way we want.
Unfortunately, right now the player and enemy react to getting shot by being pushed backwards by the bullet. I think if they are not physics objects this won't be a problem. But the player movement in the demo was implemented using physics so its a bit of work to remove that, if we want to.